pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Joaotrg09 <joao.lima@targitbrasil.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = { extras = ["standard"], version = "^0.115.13" }
uvicorn = "^0.34.3"
sqlalchemy = "^2.0.41"
psycopg2 = "^2.9.10"
jwt = "^1.4.0"
python-jose = "^3.5.0"

[tool.poetry.group.dev.dependencies]
alembic = "^1.16.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
